requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
asyncio>=3.4.3

# Google Gemini AI
google-generativeai>=0.3.0

# Web automation and browser control
browser-use>=0.1.0  # AI-powered browser automation
playwright>=1.39.0  # Required for browser-use
selenium>=4.15.2
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2

# Data processing
pandas>=2.1.0

# LangChain for AI workflows
langchain>=0.0.340
langchain-google-genai>=0.0.5

# Data validation and models
pydantic>=2.4.2
